mergetool-lib: Use $XDG_CURRENT_DESKTOP to check GNOME
To list merge tool candidates we used to use a private GNOME env
variable (GNOME_DESKTOP_SESSION_ID) that has been deprecated for long time ago
and removed as part of GNOME 3.30.0 release.

So, git should instead check the XDG_CURRENT_DESKTOP env variable, that
is supported by all the desktop environments.

Since the variable is actually a colon-separated list of names that the current
desktop is known as, we need to go through all the values to ensure
we're using GNOME